NumberedLines.jl

NumberedLines.NumberedLine

NumberedLines.attach_line_numbers

NumberedLines.detach_line_numbers

NumberedLines.with_numbered_lines

NumberedLines.without_line_number

A set of tools for easier manipulation of line number information. It can be difficult to reason about line number information when writing macros. However, it is essential to retain line number information to maintain coverage statistics. Using this package, line numbers can be directly attached to their corresponding lines. This prevents errors and makes argument counting more intuitive.

NumberedLines.NumberedLine — Type.

immutable NumberedLine

A line of code with an attached line number

julia> using NumberedLines

julia> Expr(:block, NumberedLine( LineNumberNode(1), 2) )
quote  # line 1:
    2
end

source

NumberedLines.attach_line_numbers — Method.

attach_line_numbers(a)

Attach line numbers to lines

julia> using NumberedLines

julia> e = quote
           1
           begin
               2
           end
       end |> attach_line_numbers;

julia> e.args[1].line
1

julia> e.args[2].line.args[1].line
2

julia> e == attach_line_numbers(e)
true

julia> no_line = Expr(:block, LineNumberNode(1) );

julia> Test.@test_throws ErrorException attach_line_numbers(no_line);

source

NumberedLines.detach_line_numbers — Method.

detach_line_numbers(a)

Detach line numbers from lines

julia> using NumberedLines

julia> e = quote
           1
       end;

julia> push!(e.args, 2);

julia> e |> attach_line_numbers |> detach_line_numbers == e
true

julia> detach_line_numbers(1)
1

source

NumberedLines.with_numbered_lines — Method.

with_numbered_lines(f, e)

attach_line_numbers to e, apply f, then detach_line_numbers

julia> using NumberedLines

julia> reverse_lines(e) = Expr(e.head, e.args[2], e.args[1]);

julia> e = quote
           1
           2
       end;

julia> result = with_numbered_lines(reverse_lines, e);

julia> (result.args[2], result.args[4] ) == (2, 1)
true

source

NumberedLines.without_line_number — Method.

without_line_number(f, n::NumberedLine)

Return a copy of n with f applied to the line in n

julia> using NumberedLines

julia> add_one(a) = :($a + 1);

julia> e = quote
           1
       end |> attach_line_numbers;

julia> test_line = e.args[1];

julia> without_line_number(add_one, test_line) ==
           NumberedLine(test_line.number, add_one(test_line.line) )
true

julia> without_line_number(add_one, 1) == add_one(1)
true

source